Refactor Role and Authorization Models
