+++ src/datastores/sql/crud/file.py
@@ -20,10 +20,9 @@ from sqlalchemy.orm import Session
 
 from api.v1 import schemas
 from datastores.sql.models.file import File, FileReport, FileSummary
-from datastores.sql.models.roles import Role
+from datastores.sql.models.role import Role
 
 from .folder import get_folder_from_db
-from .user import get_user_from_db
 
 
 def get_files_from_db(db: Session, folder_id: int):
